pyproject.toml
[tool.poetry]
name = "fastapi-to-openapi"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Sion908 <sionn908@gmail.com>"]
maintainers = ["Sion908 <sionn908@gmail.com>"]
readme = "README.md"
repository = "https://github.com/sion908/fastapi_to_openapi"
keywords = ["fastapi"]
classifiers = []
packages = []
include = []
exclude = []

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
fastapi = "^0.104.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fastapi_to_openapi = "fastapi_to_openapi.main:cli"

[tool.pytest.ini_options]
pythonpath = "fastapi_to_openapi"
testpaths = ["tests",]
filterwarnings = ["ignore::DeprecationWarning",]
